crc: Move tests to new crc-test.cc and fix the code on 64-bit platforms.
This code has been broken on 64-bit forever but I never noticed because
I forgot to run crc.exe as part of "make check".  After adding to
unit-tests.cc, I saw it did not work because the code was basically
assuming that 'unsigned long' is 32 bits.  Fixed by using 'uint32_t'.

+++ datablok.cc
@@ -290,7 +290,7 @@ void DataBlock::print(char const *label, int bytesPerLine) const
   if (label) {
     printf("---- %s, length = %zu, crc32 = 0x%lX ---- {\n",
            label, getDataLen(),
-           crc32(getDataC(), getDataLen()));
+           (unsigned long)crc32(getDataC(), getDataLen()));
   }
 
   size_t cursor = 0;
